refactor: improve repo local dev experience
