+++ .eslintrc
@@ -1,1 +1,6 @@
-{ "extends": "react-app" }
+{
+  "extends": "react-app",
+  "rules": {
+    "import/no-anonymous-default-export": 0
+  }
+}
