+++ .npmignore
@@ -1,14 +1,20 @@
-node_modules
+.github
+config
+schemas
 scripts
 src
-.github
+static
+super-pane
+temp-movie-project
 .eslintrc
 .gitignore
-.prettierrc
+.npmignore
+.npmrc
+.pretterrc
 .releaserc.json
 external.d.ts
-.env
-package-lock.json
-renovate.json
+rennovate.json
 rollup.config.js
+sanity.json
+structure.js
 tsconfig.json
