+++ config/@sanity/form-builder.json
