--- example-project/.eslintrc
@@ -1,1 +1,0 @@
-{ "extends": "react-app" }
