--- example-project/package.json
@@ -1,45 +1,0 @@
-{
-  "name": "exampleproject",
-  "private": true,
-  "version": "1.0.0",
-  "description": "",
-  "main": "package.json",
-  "author": "Rico Kahler <rico@bloomscape.com>",
-  "license": "UNLICENSED",
-  "scripts": {
-    "start": "sanity start",
-    "build": "sanity build"
-  },
-  "keywords": [
-    "sanity"
-  ],
-  "dependencies": {
-    "@sanity/base": "2.13.1",
-    "@sanity/components": "2.13.0",
-    "@sanity/core": "2.13.1",
-    "@sanity/default-layout": "2.13.1",
-    "@sanity/default-login": "2.13.1",
-    "@sanity/desk-tool": "2.13.1",
-    "@sanity/google-maps-input": "2.13.1",
-    "@sanity/vision": "2.13.1",
-    "@types/classnames": "2.3.0",
-    "@types/react-window": "1.8.5",
-    "@typescript-eslint/eslint-plugin": "4.28.5",
-    "@typescript-eslint/parser": "4.28.5",
-    "babel-eslint": "10.1.0",
-    "eslint": "7.32.0",
-    "eslint-config-react-app": "6.0.0",
-    "eslint-plugin-flowtype": "5.8.2",
-    "eslint-plugin-import": "2.23.4",
-    "eslint-plugin-jsx-a11y": "6.4.1",
-    "eslint-plugin-react": "7.24.0",
-    "eslint-plugin-react-hooks": "4.2.0",
-    "prop-types": "15.7.2",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-error-boundary": "3.1.3",
-    "react-icons": "4.2.0",
-    "react-window": "1.8.6",
-    "typescript": "4.3.5"
-  }
-}
