--- example-project/plugins/.gitkeep
@@ -1,1 +1,0 @@
-User-specific packages can be placed here
