--- example-project/tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "compilerOptions": {
-    "strict": true,
-    "jsx": "preserve",
-    "esModuleInterop": true,
-    "target": "esnext",
-    "moduleResolution": "node"
-  },
-  // Note: This config is only used to help editors like VS Code understand/resolve
-  // parts, the actual transpilation is done by babel. Any compiler configuration in
-  // here will be ignored.
-  "include": [
-    "./node_modules/@sanity/base/types/**/*.ts",
-    "./**/*.ts",
-    "./**/*.tsx"
-  ]
-}
