+++ package.json
@@ -14,10 +14,11 @@
   "module": "./dist/index.esm.js",
   "types": "./dist/index.d.ts",
   "scripts": {
+    "start": "sanity start",
     "test": "echo \"Error: no test specified\" && exit 1",
     "build": "./scripts/build",
     "prepare": "npm run build",
-    "lint": "eslint src",
+    "lint": "eslint super-pane",
     "format": "prettier . --write"
   },
   "dependencies": {
@@ -36,9 +37,18 @@
     "@babel/preset-typescript": "7.14.5",
     "@rollup/plugin-babel": "5.3.0",
     "@rollup/plugin-node-resolve": "11.2.1",
+    "@sanity/base": "2.13.1",
     "@sanity/client": "2.13.0",
+    "@sanity/components": "2.13.0",
+    "@sanity/core": "2.13.1",
+    "@sanity/default-layout": "2.13.1",
+    "@sanity/default-login": "2.13.1",
+    "@sanity/desk-tool": "2.13.1",
+    "@sanity/google-maps-input": "2.13.1",
     "@sanity/icons": "1.1.3",
     "@sanity/ui": "0.34.7",
+    "@sanity/vision": "2.13.1",
+    "@types/classnames": "2.3.0",
     "@types/react": "17.0.15",
     "@types/react-dom": "17.0.9",
     "@types/styled-components": "5.1.11",
@@ -53,10 +63,14 @@
     "eslint-plugin-jsx-a11y": "6.4.1",
     "eslint-plugin-react": "7.24.0",
     "eslint-plugin-react-hooks": "4.2.0",
+    "json-reduce": "^2.0.0",
     "postcss": "8.3.6",
     "prettier": "2.3.2",
+    "prop-types": "15.7.2",
     "react": "17.0.2",
     "react-dom": "17.0.2",
+    "react-error-boundary": "3.1.3",
+    "react-icons": "4.2.0",
     "rollup": "2.55.1",
     "rollup-plugin-postcss": "4.0.0",
     "rxjs": "6.6.7",
@@ -66,6 +80,8 @@
   "peerDependencies": {
     "@sanity/icons": "^1.0.0",
     "@sanity/ui": "*",
+    "@sanity/desk-tool": "*",
+    "@sanity/base": "*",
     "@types/react": "^16 || ^17",
     "react": "^16 || ^17",
     "rxjs": "^6.0.0"
