+++ super-pane/block-content-to-string.ts
