+++ super-pane/bulk-actions-menu/styles.module.css
