+++ super-pane/create-emitter.ts
