+++ super-pane/styles.module.css
